@@ -1,5 +1,8 @@
 // Error response in development environment
-const sendErrorDev = (err, res) => {
+
+import AppError from 'src/utils/appError';
+
+const sendErrorDev = (err: AppError, res): void => {
   res.status(err.statusCode).json({
     status: err.status,
     error: err,
@@ -9,15 +12,13 @@ const sendErrorDev = (err, res) => {
 };
 
 // Error response in production environment
-const sendErrorProd = (err, res) => {
+const sendErrorProd = (err: AppError, res): void => {
   if (err.isOperational) {
     res.status(err.statusCode).json({
       status: err.status,
       message: err.message,
     });
   } else {
-    console.error('ERROR', err);
-
     res.status(500).json({
       status: 'error',
       message: 'Something went very wrong',
@@ -25,15 +26,15 @@ const sendErrorProd = (err, res) => {
   }
 };
 
-module.exports = (err, req, res, next) => {
+export default (err: AppError, req, res): void => {
   err.statusCode = err.statusCode || 500;
   err.status = err.status || 'error';
 
-  if (process.env.NODE_ENV === 'development') {
-    sendErrorDev(err, res);
-  } else if (process.env.NODE_ENV === 'production') {
+  if (process.env.NODE_ENV === 'production') {
     // Configure other error types
 
     sendErrorProd(err, res);
+  } else {
+    sendErrorDev(err, res);
   }
 };
